Updated generated code for Linux Updated issue with postcript. Fixed issue with the example 04.

+++ library/post_script.sh
@@ -10,26 +10,34 @@
 #cp Makefile.SH  ./generated_wrapper/c/src                          --example
 cd generated_wrapper/c/include
 mkdir linux 
-cp ewg_sdl2_callback_c_glue_code.h .\linux
+cp ewg_sdl2_callback_c_glue_code.h ./linux
 rm ewg_sdl2_callback_c_glue_code.h
-cd ..\src
+cd ..
+cd src
+
 mkdir linux 
-cp build.eant .\linux
+cp build.eant ./linux
 rm build.eant
-cp ewg_sdl2_callback_c_glue_code.c .\linux
+cp ewg_sdl2_callback_c_glue_code.c ./linux
 rm ewg_sdl2_callback_c_glue_code.c
 rm Makefile.SH  
+
 cd ..
 cd ..
 cd ..
-cp Makefile.SH  .\generated_wrapper\c\src\linux
+cp Makefile.SH  ./generated_wrapper/c/src/linux
 
 cd generated_wrapper/eiffel/
 rm  sdl_constant_api.e
-rem sdl_event_union_api.e
-mkdir linux 
-cp ewg_sdl2_callback_c_glue_code_functions_api.e .\linux
-rm ewg_sdl2_callback_c_glue_code_functions_api.e 
+rm  sdl_event_union_api.e
+
+mkdir linux
+FILE=ewg_sdl2_callback_c_glue_code_functions_api.e
+if test -f "$FILE"; then
+    cp ewg_sdl2_callback_c_glue_code_functions_api.e ./linux
+    rm ewg_sdl2_callback_c_glue_code_functions_api.e 
+fi
+ 
 
 cd ..
 cd ..
@@ -39,5 +47,5 @@ cd ..
 
 
 
-cd generated_wrapper/c/src/
+cd generated_wrapper/c/src/linux
 finish_freezing -library
